Removing EventSource. The class was used to register listeners EventListeners to HornetQ so that th listeners could receive messages from HornetQ. Each listener has been wrapped in ListenerWrapper.
Removing ListenerWrapper. The class was a HornetQ MessageHandler implementation. It was in charge of deserialization of messages that came out of HornetQ and delivering it to the actual Candlepin EventListener implementation.

Renamed HornetqContextListener to EventingConfigurationContextListener. The class used to start EmbeddedHornetQ server. It was also configuring so called HornetQ listeners, which still needs to be done. Now we should better call them just Event Listeners.

Deleting most of the EventSink implementation. Up to know it used to queue up events in a transaction. When sendEvents() was called, it commited the tx.