Suspend Mode
Candlepin can enter Suspend Mode. In this mode Candlepin will only respond to
Status Resource requests and pause all Pinsetter Jobs. It is expected that as
soon as Candlepin enters Suspend Mode, no more new requests or jobs are started.
It expected that current requests and jobs will be terminated as soon as
possible. To do so, various common parts of logic such as EventSink will query
ModeManager for current mode and fail fast. It is still possible that some
of the ongoing requests and jobs succeed.